+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ HEAD
   - fix issue with pending tags not being sorted ;)
   - fix the issue with incorrectly flattened values in query
   - extend the memory collector to include the allocation order as resource
+  - add loading and storing of performance change records
 
 0.14.1 (2018-04-19)
 -------------------